Change package name from `hgl-rs` to `hgl`
When the Cargo package manager gets the online package list component, the package name 
will be used as the identifier for the whole project. Keeping "-rs" -- while useful for the github 
repository -- will be redundant in Cargo because all of the projects will obviously be rust projects.

This rational was used to justify the same change in `gl-rs`.

This patch also fixes the dependency on `gl-rs` to use the new package name.